Update VolatilityDataSet.java
Will hold Volatility data for all 30 companies on DJIA. Then this bot will pick the top 5 or 10 stocks in terms of volatility and start the trading routine on all of them. Thinking of using a parallelized quicksort to find the most volatile stocks, but given the small size (30) elements, it might be overkill...lol.